src/streamlit/requirements.txt
streamlit
pandas
plotly.express
sqlalchemy
psycopg2
psycopg2-binary